Added dokan chainable container for background process and refactored upgrader to there
